# Upgrade the ProtocolBuffer compiler.