--- a/Rakefile
+++ b/Rakefile
@@ -3,27 +3,28 @@
 #
 
 protoc_binary="protoc"
-protoc_version="3.5.1"
-gogo_dir="/tmp/gogo"
+protoc_version="21.12"
 
 namespace :codegen do
 
   task :install_protoc do
     if `bash -c "protoc --version"` != "libprotoc ${protoc_version}"
-      protoc_binary="/tmp/protoc#{protoc_version}"
+      protoc_binary="/tmp/protoc-#{protoc_version}/bin/protoc"
       sh <<-EOF
         /bin/bash <<BASH
         if [ ! -f #{protoc_binary} ] ; then
           echo "Downloading protoc #{protoc_version}"
-          cd /tmp
+          rm -rf /tmp/protoc-#{protoc_version}
+          mkdir /tmp/protoc-#{protoc_version}
+          cd /tmp/protoc-#{protoc_version}
           if [ "$(uname -s)" = "Darwin" ] ; then
-            curl -OL https://github.com/google/protobuf/releases/download/v#{protoc_version}/protoc-#{protoc_version}-osx-x86_64.zip
+            curl -OL https://github.com/protocolbuffers/protobuf/releases/download/v#{protoc_version}/protoc-#{protoc_version}-osx-x86_64.zip
           else
-            curl -OL https://github.com/google/protobuf/releases/download/v#{protoc_version}/protoc-#{protoc_version}-linux-x86_64.zip
+            curl -OL https://github.com/protocolbuffers/protobuf/releases/download/v#{protoc_version}/protoc-#{protoc_version}-linux-x86_64.zip
           fi
           unzip protoc-#{protoc_version}*.zip
-          mv bin/protoc #{protoc_binary}
         fi
+        go install google.golang.org/protobuf/cmd/protoc-gen-go@latest
 BASH
       EOF
     end
@@ -36,40 +37,20 @@ BASH
 
       export GO111MODULE=auto
 
-      rm -rf #{gogo_dir}
-      rm -rf /tmp/gogo-bin-*
-
-      mkdir -p #{gogo_dir}/src/github.com/gogo
-      git clone https://github.com/gogo/protobuf.git #{gogo_dir}/src/github.com/gogo/protobuf
-
-      # Install v1.0.0
-      pushd #{gogo_dir}/src/github.com/gogo/protobuf
-      git checkout v1.0.0
-      GOBIN=/tmp/gogo-bin-v1.0.0 GOPATH=#{gogo_dir} make clean install
-
-      popd
-
       echo "Generating logs proto"
-      PATH=/tmp/gogo-bin-v1.0.0 #{protoc_binary} --proto_path=$GOPATH/src:#{gogo_dir}/src:. --gogofast_out=$GOPATH/src --java_out=java proto/logs/agent_logs_payload.proto
+      PATH=$GOPATH/bin /tmp/protoc-#{protoc_version}/bin/protoc --proto_path=$GOPATH/src:. --go_out=$GOPATH/src --java_out=java proto/logs/agent_logs_payload.proto
 
       echo "Generating metrics proto (go)"
-      PATH=/tmp/gogo-bin-v1.0.0 #{protoc_binary} --proto_path=$GOPATH/src:#{gogo_dir}/src:. --gogofast_out=$GOPATH/src proto/metrics/agent_payload.proto
+      PATH=$GOPATH/bin /tmp/protoc-#{protoc_version}/bin/protoc --proto_path=$GOPATH/src:. --go_out=$GOPATH/src proto/metrics/agent_payload.proto
 
       echo "Generating metrics proto (python)"
-      #{protoc_binary} --proto_path=#{gogo_dir}/src:$GOPATH/src:./proto/metrics --python_out=python agent_payload.proto
-
-      # Install the specific tag that the process-agent needs
-      pushd #{gogo_dir}/src/github.com/gogo/protobuf
-      git checkout d76fbc1373015ced59b43ac267f28d546b955683
-      GOBIN=/tmp/gogo-bin-d76fbc1373015ced59b43ac267f28d546b955683 GOPATH=#{gogo_dir} make clean install
-
-      popd
+      PATH=$GOPATH/bin /tmp/protoc-#{protoc_version}/bin/protoc --proto_path=$GOPATH/src:./proto/metrics --python_out=python agent_payload.proto
 
       echo "Generating process proto"
-      PATH=/tmp/gogo-bin-d76fbc1373015ced59b43ac267f28d546b955683 #{protoc_binary} --proto_path=$GOPATH/src:#{gogo_dir}/src:. --gogofaster_out=$GOPATH/src proto/process/*.proto
+      PATH=$GOPATH/bin /tmp/protoc-#{protoc_version}/bin/protoc --proto_path=$GOPATH/src:. --go_out=$GOPATH/src proto/process/*.proto
 
       echo "Generating contlcycle proto"
-      PATH=/tmp/gogo-bin-v1.0.0 #{protoc_binary} --proto_path=$GOPATH/src:#{gogo_dir}/src:. --gogofast_out=$GOPATH/src proto/contlcycle/contlcycle.proto
+      PATH=$GOPATH/bin /tmp/protoc-#{protoc_version}/bin/protoc --proto_path=$GOPATH/src:. --go_out=$GOPATH/src proto/contlcycle/contlcycle.proto
 
       cp -r v5/* .
       rm -rf v5