--- a/go.mod
+++ b/go.mod
@@ -6,11 +6,11 @@ require (
 	github.com/DataDog/mmh3 v0.0.0-20200805151601-30884ca2197a
 	github.com/DataDog/zstd v1.4.8
 	github.com/DataDog/zstd_0 v0.0.0-20210310093942-586c1286621f
-	github.com/gogo/protobuf v1.0.0
 	github.com/stretchr/testify v1.3.0
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	google.golang.org/protobuf v1.28.1 // indirect
 )