--- a/java/com/dd/agent/pb/AgentPayload.java
+++ b/java/com/dd/agent/pb/AgentPayload.java
@@ -36,18 +36,10 @@ public final class AgentPayload {
       "Z&github.com/DataDog/agent-payload/v5/pb" +
       "b\006proto3"
     };
-    com.google.protobuf.Descriptors.FileDescriptor.InternalDescriptorAssigner assigner =
-        new com.google.protobuf.Descriptors.FileDescriptor.    InternalDescriptorAssigner() {
-          public com.google.protobuf.ExtensionRegistry assignDescriptors(
-              com.google.protobuf.Descriptors.FileDescriptor root) {
-            descriptor = root;
-            return null;
-          }
-        };
-    com.google.protobuf.Descriptors.FileDescriptor
+    descriptor = com.google.protobuf.Descriptors.FileDescriptor
       .internalBuildGeneratedFileFrom(descriptorData,
         new com.google.protobuf.Descriptors.FileDescriptor[] {
-        }, assigner);
+        });
     internal_static_pb_Log_descriptor =
       getDescriptor().getMessageTypes().get(0);
     internal_static_pb_Log_fieldAccessorTable = new