--- a/proto/metrics/agent_payload.proto
+++ b/proto/metrics/agent_payload.proto
@@ -2,8 +2,6 @@ syntax = "proto3";
 
 package datadog.agentpayload;
 
-import "github.com/gogo/protobuf/gogoproto/gogo.proto";
-
 option go_package = "github.com/DataDog/agent-payload/v5/gogen";
 
 message CommonMetadata {
@@ -101,12 +99,12 @@ message SketchPayload {
 		}
 		string metric = 1;
 		string host = 2;
-		repeated Distribution distributions = 3 [(gogoproto.nullable) = false];
+		repeated Distribution distributions = 3;
 		repeated string tags = 4;
                 reserved 5, 6;
 				reserved "distributionsK", "distributionsC";
-		repeated Dogsketch dogsketches = 7 [(gogoproto.nullable) = false];
+		repeated Dogsketch dogsketches = 7;
 	}
-	repeated Sketch sketches = 1 [(gogoproto.nullable) = false];
-	CommonMetadata metadata = 2 [(gogoproto.nullable) = false];
+	repeated Sketch sketches = 1;
+	CommonMetadata metadata = 2;
 }