… contlcycle/contlcycle.pb.go
… go.sum
… gogen/agent_payload.pb.go
… java/com/dd/agent/pb/Log.java
… pb/agent_logs_payload.pb.go
… process/agent.pb.go
… process/containeraddr.go
… process/events.pb.go
… python/agent_payload_pb2.py