@L3n41c 2022-12-23
Here are the changes needed on `datadog-agent` side to use this new version: DataDog/datadog-agent#14844.